src/main/java/tasks/leetcode/hashmap/ContainsDuplicateII.java
package tasks.leetcode.hashmap;

import java.util.HashMap;
import java.util.Map;

/**
 * === Мапа и расстояние между значениями по индексу ===
 */
public class ContainsDuplicateII {
    public static boolean containsNearbyDuplicate(int[] nums, int k) {
        Map<Integer, Integer> map = new HashMap<>();
        for (int i = 0; i < nums.length; i++) {
            if (map.containsKey(nums[i])) {
                if (i - map.get(nums[i]) <= k) {
                    return true;
                }
            }
            map.put(nums[i], i);
        }
        return false;
    }
}
